Refactoring attempt 2
